methods/unicorn/requirements.txt
pathtype==1.0.0
numpy==1.20.3
pandas
scipy==1.5.0
transformers==4.6.1
matplotlib==3.2.2
torchvision==0.8.2
torchaudio==0.7.2
